Created Add-BMObjectParameter function.
Many of the BuildMaster APIs accept IDs or names to identify what objects to retrieve. Add-BMObjectParameter will allows BuildMasterAutomation functions to accept IDs, names, or actual objects with ID and name parameters for parameters for these identity keys. Add-BMObjectParameter parses objects to determine what the appropriate key is to send to the BuildMaster API.